backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
scikit-learn
requests
python-dotenv
SQLAlchemy>=2.0
celery>=5.3.0
redis>=5.0.0
python-multipart>=0.0.9
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
razorpay
